refactor: ts migration for idv-submit-complete
