+++ packages/account/src/Components/poi/idv-status/idv-submit-complete/idv-submit-complete.tsx
@@ -4,7 +4,12 @@ import { localize } from '@deriv/translations';
 import PoaButton from 'Components/poa/poa-button';
 import IdvDocumentPending from 'Assets/ic-idv-document-pending.svg';
 
-const IdvSubmitComplete = ({ needs_poa, is_from_external }) => {
+type TIdvSubmitComplete = {
+    needs_poa: boolean;
+    is_from_external: boolean;
+};
+
+const IdvSubmitComplete = ({ needs_poa, is_from_external }: TIdvSubmitComplete) => {
     const poa_button = !is_from_external && <PoaButton custom_text={localize('Submit proof of address')} />;
 
     return (
