+++ packages/account/src/Components/poi/idv-status/idv-submit-complete/index.js
@@ -1,3 +1,3 @@
-import IdvSubmitComplete from './idv-submit-complete.jsx';
+import IdvSubmitComplete from './idv-submit-complete';
 
 export default IdvSubmitComplete;
