+++ packages/account/tsconfig.json
@@ -4,6 +4,7 @@
         "outDir": "./dist",
         "baseUrl": "./",
         "paths": {
+            "Assets/*": ["src/Assets/*"],
             "Components/*": ["src/Components/*"],
             "Containers/*": ["src/Containers/*"],
             "Constants/*": ["src/Constants/*"],
